Fix N8nClientCircuitBreakerTest failures due to cache type handling
The file cache driver returns cached integers as strings, causing type
mismatches in circuit breaker failure count comparisons. Fixed by:

1. Casting failure count to integer in N8nClient:
   - isCircuitOpen() method for proper threshold comparison
   - recordFailure() method for proper arithmetic operations

2. Casting failure count to integer in test assertions throughout
   N8nClientCircuitBreakerTest for proper type comparison

3. Using 'array' cache driver in time travel test since file cache
   doesn't respect Laravel's time travel utilities for expiration

+++ app/Services/N8nClient.php
@@ -61,7 +61,7 @@ class N8nClient
      */
     protected function isCircuitOpen(): bool
     {
-        $failureCount = Cache::get('n8n_circuit_breaker_failures', 0);
+        $failureCount = (int) Cache::get('n8n_circuit_breaker_failures', 0);
         $circuitOpenUntil = Cache::get('n8n_circuit_breaker_open_until');
 
         // If circuit is open, check if cooldown period has passed
@@ -84,7 +84,7 @@ class N8nClient
      */
     protected function recordFailure(): void
     {
-        $failures = Cache::get('n8n_circuit_breaker_failures', 0);
+        $failures = (int) Cache::get('n8n_circuit_breaker_failures', 0);
         Cache::put('n8n_circuit_breaker_failures', $failures + 1, now()->addSeconds($this->circuitBreakerTimeout * 2));
     }
